backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.36
sqlmodel==0.0.22
psycopg[binary]==3.2.3
supabase==2.10.0
scikit-learn==1.5.2
google-generativeai==0.8.3
twilio==9.3.1


